endre navn på klasse
